Make upload workers faster on processing layers
Make upload workers processing image layers only once (as the best
effort) and avoid re-fetching it from a registry.

Whether cross-linking that known layer for another local images, or
performing cross-repo blobs patching, leverage the cached in memory
global view shared for all workers to speed-up processing of that
already processed layer.

The 1st time a layer gets uploaded for an image, that image and
its known path (local or remote) becomes a reference for future
cross-referencing by other images.

Having that global view implemented skip uploading the 1st image in
the tasks list as a separate (and non-concurrent) job.

Based on the dynamically picked multi-workers mode, provide the global
view graf state synchronization as the following:

* if it cannot use multi-process workers, like when executing it via
  Mistral by monkey patched eventlet greenthreads, choose threadinglock
  and multi-threads-safe standard dictionary in the shared class
  namespace to store the global view there
* if it can do MP, pick processlock and its (safe from data races)
  Manager().dict() as the global view shared among cooperating OS
  processes.
* use that global view in a transparent fashion, provided by a special
  classmethod proxying access to the internal state shared for workers.

Ultimately, all that optimizes:

* completion time
* re-fetching of the already processed layers
* local deduplication of layers
* the amount of outbound HTTP requests to registries
* if-layer-exists and other internal logic check executed against the
  in-memory cache firstly.

As layers locking and unlocking becomes a popular action, reduce the
noise of the debug messages it produces.

Closes-bug: #1847225
Related-bug: #1844446

Change-Id: Ie5ef4045b7e22c06551e886f9f9b6f22c8d4bd21
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>